+++ crates/ruma-events/src/kinds.rs
@@ -17,6 +17,8 @@ use super::{
     RedactionDeHelper, RoomAccountDataEventContent, StateEventType, StaticStateEventContent,
     ToDeviceEventContent,
 };
+#[cfg(feature = "unstable-msc4354")]
+use crate::sticky::StickyObject;
 
 /// A global account data event.
 #[derive(Clone, Debug, Event)]
@@ -167,6 +169,13 @@ pub struct OriginalMessageLikeEvent<C: MessageLikeEventContent> {
 
     /// Additional key-value pairs not signed by the homeserver.
     pub unsigned: MessageLikeUnsigned<C>,
+
+    /// Message events can be annotated with a new top-level sticky object,
+    /// which MUST have a duration_ms, which is the number of milliseconds for the event to be
+    /// sticky.
+    #[cfg(feature = "unstable-msc4354")]
+    #[ruma_event(default, default_on_error, rename = "msc4354_sticky")]
+    pub sticky: Option<StickyObject>,
 }
 
 impl<C: MessageLikeEventContent> JsonCastable<OriginalSyncMessageLikeEvent<C>>
@@ -211,6 +220,13 @@ pub struct OriginalSyncMessageLikeEvent<C: MessageLikeEventContent> {
 
     /// Additional key-value pairs not signed by the homeserver.
     pub unsigned: MessageLikeUnsigned<C>,
+
+    /// Message events can be annotated with a new top-level sticky object,
+    /// which MUST have a duration_ms, which is the number of milliseconds for the event to be
+    /// sticky.
+    #[cfg(feature = "unstable-msc4354")]
+    #[ruma_event(default, default_on_error, rename = "msc4354_sticky")]
+    pub sticky: Option<StickyObject>,
 }
 
 impl<C: MessageLikeEventContent + RedactContent> OriginalSyncMessageLikeEvent<C>
@@ -398,6 +414,13 @@ pub struct OriginalStateEvent<C: StaticStateEventContent> {
 
     /// Additional key-value pairs not signed by the homeserver.
     pub unsigned: C::Unsigned,
+
+    /// Message events can be annotated with a new top-level sticky object,
+    /// which MUST have a duration_ms, which is the number of milliseconds for the event to be
+    /// sticky.
+    #[cfg(feature = "unstable-msc4354")]
+    #[ruma_event(default, default_on_error, rename = "msc4354_sticky")]
+    pub sticky: Option<StickyObject>,
 }
 
 impl<C: StaticStateEventContent> JsonCastable<OriginalSyncStateEvent<C>> for OriginalStateEvent<C> {}
@@ -454,6 +477,13 @@ pub struct OriginalSyncStateEvent<C: StaticStateEventContent> {
 
     /// Additional key-value pairs not signed by the homeserver.
     pub unsigned: C::Unsigned,
+
+    /// Message events can be annotated with a new top-level sticky object,
+    /// which MUST have a duration_ms, which is the number of milliseconds for the event to be
+    /// sticky.
+    #[cfg(feature = "unstable-msc4354")]
+    #[ruma_event(default, default_on_error, rename = "msc4354_sticky")]
+    pub sticky: Option<StickyObject>,
 }
 
 impl<C: StaticStateEventContent + RedactContent> JsonCastable<SyncStateEvent<C>>